Moved channel torrent tests
