+++ Tribler/Test/Core/Modules/RestApi/test_my_channel_endpoints.py
@@ -238,188 +238,6 @@ class TestMyChannelEndpoints(AbstractTestMyChannelEndpoints):
         return self.do_request('mychannel/rssfeeds', expected_code=404)
 
 
-class TestMyChannelTorrentsEndpoint(AbstractTestMyChannelEndpoints):
-
-    @deferred(timeout=10)
-    def test_add_torrent_to_my_channel(self):
-        my_channel_id = self.create_fake_channel("channel", "")
-        torrent_path = TORRENT_FILE
-
-        def verify_method_invocation(channel_id, torrent_def, extra_info={}, forward=True):
-            self.assertEqual(my_channel_id, channel_id)
-            self.assertEqual(TorrentDef.load(torrent_path), torrent_def)
-            self.assertEqual({}, extra_info)
-            self.assertEqual(True, forward)
-
-        self.session.add_torrent_def_to_channel = verify_method_invocation
-
-        torrent_file = open(torrent_path, mode='rb')
-        torrent_64 = base64.b64encode(torrent_file.read())
-        post_data = {
-            "torrent": torrent_64
-        }
-        expected_json = {"added": True}
-        return self.do_request('mychannel/torrents', 200, expected_json, 'PUT', post_data)
-
-    @deferred(timeout=10)
-    def test_add_torrent_to_my_channel_with_description(self):
-        my_channel_id = self.create_fake_channel("channel", "")
-        torrent_path = TORRENT_FILE
-
-        def verify_method_invocation(channel_id, torrent_def, extra_info={}, forward=True):
-            self.assertEqual(my_channel_id, channel_id)
-            self.assertEqual(TorrentDef.load(torrent_path), torrent_def)
-            self.assertEqual({"description": "video of my cat"}, extra_info)
-            self.assertEqual(True, forward)
-
-        self.session.add_torrent_def_to_channel = verify_method_invocation
-
-        torrent_file = open(torrent_path, mode='rb')
-        torrent_64 = base64.b64encode(torrent_file.read())
-        post_data = {
-            "torrent": torrent_64,
-            "description": "video of my cat"
-        }
-        expected_json = {"added": True}
-        return self.do_request('mychannel/torrents', 200, expected_json, 'PUT', post_data)
-
-    @deferred(timeout=10)
-    def test_add_torrent_to_my_channel_404(self):
-        return self.do_request('mychannel/torrents', 404, None, 'PUT')
-
-    @deferred(timeout=10)
-    def test_add_torrent_to_my_channel_missing_parameter(self):
-        self.create_fake_channel("channel", "")
-        expected_json = {"error": "torrent parameter missing"}
-        return self.do_request('mychannel/torrents', 400, expected_json, 'PUT')
-
-    @deferred(timeout=10)
-    def test_add_torrent_to_my_channel_500(self):
-        """
-        Testing whether the API returns a formatted 500 error if ValueError is raised
-        """
-        self.create_fake_channel("channel", "")
-        torrent_path = TORRENT_FILE
-
-        def fake_error(channel_id, torrent_def, extra_info={}, forward=True):
-            raise ValueError("Test error")
-
-        self.session.add_torrent_def_to_channel = fake_error
-
-        def verify_error_message(body):
-            error_response = json.loads(body)
-            expected_response = {
-                u"error": {
-                    u"handled": True,
-                    u"code": u"ValueError",
-                    u"message": u"Test error"
-                }
-            }
-            self.assertDictContainsSubset(expected_response[u"error"], error_response[u"error"])
-
-        torrent_file = open(torrent_path, mode='rb')
-        torrent_64 = base64.b64encode(torrent_file.read())
-        post_data = {
-            "torrent": torrent_64
-        }
-        self.should_check_equality = False
-        return self.do_request('mychannel/torrents', expected_code=500, expected_json=None, request_type='PUT',
-                               post_data=post_data).addCallback(verify_error_message)
-
-
-class TestMyChannelModifyTorrentsEndpoint(AbstractTestMyChannelEndpoints):
-
-    def setUpPreSession(self):
-        super(TestMyChannelModifyTorrentsEndpoint, self).setUpPreSession()
-        self.config.set_libtorrent(True)
-
-    def test_add_torrent_from_url_to_my_channel_with_description(self):
-        """
-        Testing whether a torrent can be added to a channel using the API
-        """
-        my_channel_id = self.create_fake_channel("channel", "")
-
-        # Setup file server to serve torrent file
-        files_path = os.path.join(self.session_base_dir, 'http_torrent_files')
-        os.mkdir(files_path)
-        shutil.copyfile(TORRENT_FILE, os.path.join(files_path, 'ubuntu.torrent'))
-        file_server_port = get_random_port()
-        self.setUpFileServer(file_server_port, files_path)
-
-        def verify_method_invocation(channel_id, torrent_def, extra_info={}, forward=True):
-            self.assertEqual(my_channel_id, channel_id)
-            self.assertEqual(TorrentDef.load(os.path.join(files_path, 'ubuntu.torrent')), torrent_def)
-            self.assertEqual({}, extra_info)
-            self.assertEqual(True, forward)
-
-        self.session.add_torrent_def_to_channel = verify_method_invocation
-
-        # Martijn: We are directly invoking the endpoint here. This is because the requests library used by
-        # TorrentDef.load_from_url is not working well when performing a request using do_request and seems to hang.
-        torrent_url = 'http://localhost:%s/ubuntu.torrent' % file_server_port
-        endpoint = MyChannelModifyTorrentsEndpoint(self.session, torrent_url)
-        fake_request = MockObject()
-        fake_request.content = MockObject()
-        fake_request.content.read = lambda: ""
-        self.assertDictEqual(json.loads(endpoint.render_PUT(fake_request)), {"added": torrent_url})
-
-    @deferred(timeout=10)
-    def test_add_torrent_from_magnet_to_my_channel_without_description(self):
-        my_channel_id = self.create_fake_channel("channel", "")
-        torrent_path = TORRENT_FILE
-
-        def fake_load_from_dht(infohash_or_magnet, callback):
-            meta_info = TorrentDef.load(torrent_path).get_metainfo()
-            callback(meta_info)
-
-        self.session.lm.ltmgr.get_metainfo = fake_load_from_dht
-
-        def verify_method_invocation(channel_id, torrent_def, extra_info={}, forward=True):
-            self.assertEqual(my_channel_id, channel_id)
-            self.assertEqual(TorrentDef.load(torrent_path), torrent_def)
-            self.assertEqual({}, extra_info)
-            self.assertEqual(True, forward)
-
-        self.session.add_torrent_def_to_channel = verify_method_invocation
-
-        magnet_url = 'magnet:?fake'
-        url = 'mychannel/torrents/' + urllib.quote_plus(magnet_url)
-        return self.do_request(url, expected_code=200, expected_json={"added": magnet_url}, request_type='PUT')
-
-    @deferred(timeout=10)
-    def test_add_torrent_to_my_channel_404(self):
-        return self.do_request('mychannel/torrents/fake_url', expected_code=404, expected_json=None, request_type='PUT')
-
-    @deferred(timeout=10)
-    def test_add_torrent_to_my_channel_500(self):
-        """
-        Testing whether the API returns a formatted 500 error if ValueError is raised
-        """
-        self.create_fake_channel("channel", "")
-
-        def fake_get_metainfo(infohash_or_magnet, callback):
-            raise ValueError(u"Test error")
-
-        self.session.lm.ltmgr.get_metainfo = fake_get_metainfo
-
-        def verify_error_message(body):
-            error_response = json.loads(body)
-            expected_response = {
-                u"error": {
-                    u"handled": True,
-                    u"code": u"ValueError",
-                    u"message": u"Test error"
-                }
-            }
-            self.assertDictContainsSubset(expected_response[u"error"], error_response[u"error"])
-
-        torrent_url = 'magnet:testtest'
-        url = 'mychannel/torrents/' + urllib.quote_plus(torrent_url)
-        self.should_check_equality = False
-        return self.do_request(url, expected_code=500, expected_json=None, request_type='PUT')\
-                   .addCallback(verify_error_message)
-
-
 class TestMyChannelRssEndpoints(AbstractTestMyChannelEndpoints):
 
     @deferred(timeout=10)
